requirements.txt
streamlit==1.38.0
spacy==3.7.5
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.0.1
nltk==3.8.1
python-docx==1.1.2
pandas==2.2.2
markdown==3.7
google-api-python-client==2.150.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
scikit-learn==1.5.1
numpy==2.1.1